# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Bank Cards API
  description: Система управления банковскими картами
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Generated server url
...
paths:
  "/api/users":
    get:
      tags:
        - Пользователи
      summary: Получить всех пользователей (только ADMIN)
      description: Возвращает список всех пользователей с пагинацией
      operationId: getAll
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            format: int32
            default: 1
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            format: int32
            default: 10
        - name: sort
          in: query
          required: false
          schema:
            type: string
            default: id
        - name: direction
          in: query
          required: false
          schema:
            type: string
            default: asc
      responses:
        '200':
          description: Список пользователей
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/PagedModelUserDto"
        '403':
          description: Доступ запрещен
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/PagedModelUserDto"
      security:
        - bearerAuth: []
    put:
      tags:
        - Пользователи
      summary: Создать пользователя (только ADMIN)
      description: Создает нового пользователя в системе
      operationId: create
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/CreateUserRequest"
        required: true
      responses:
        '200':
          description: Пользователь создан
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/UserDto"
        '400':
          description: Пользователь уже существует
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/UserDto"
        '403':
          description: Доступ запрещен
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/UserDto"
      security:
        - bearerAuth: []
...
  "/api/users/{id}/enabled":
    post:
      tags:
        - Пользователи
      summary: Включить/отключить пользователя (только ADMIN)
      description: Изменяет статус активности пользователя
      operationId: setEnabled
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: value
          in: query
          required: true
          schema:
            type: boolean
      responses:
        '200':
          description: Статус пользователя изменен
          content:
            application/json:
              schema:
                type: object
        '400':
          description: Пользователь не найден
          content:
            application/json:
              schema:
                type: object
        '403':
          description: Доступ запрещен
          content:
            application/json:
              schema:
                type: object
      security:
        - bearerAuth: []
...
  "/api/users/{id}":
    get:
      tags:
        - Пользователи
      summary: Получить пользователя по ID (только ADMIN)
      description: Возвращает информацию о конкретном пользователе
      operationId: get
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Информация о пользователе
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/UserDto"
        '400':
          description: Пользователь не найден
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/UserDto"
        '403':
          description: Доступ запрещен
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/UserDto"
      security:
        - bearerAuth: []
  "/api/users/{id}/cards":
    get:
      tags:
        - Пользователи
      summary: Получить карты пользователя (только ADMIN)
      description: Возвращает список карт указанного пользователя с пагинацией
      operationId: getUserCards
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: page
          in: query
          required: false
          schema:
            type: integer
            format: int32
            default: 1
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            format: int32
            default: 10
        - name: sort
          in: query
          required: false
          schema:
            type: string
            default: id
        - name: direction
          in: query
          required: false
          schema:
            type: string
            default: asc
      responses:
        '200':
          description: Список карт пользователя
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/PagedModelCardDto"
        '400':
          description: Пользователь не найден
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/PagedModelCardDto"
        '403':
          description: Доступ запрещен
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/PagedModelCardDto"
      security:
        - bearerAuth: []
  "/api/users/me":
    get:
      tags:
        - Пользователи
      summary: Получить информацию о текущем пользователе
      description: Возвращает информацию о текущем аутентифицированном пользователе
      operationId: me
      responses:
        '200':
          description: Информация о пользователе
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/UserDto"
      security:
        - bearerAuth: []
  "/api/transactions/user/{userId}":
    get:
      tags:
        - Транзакции
      summary: Получить транзакции по пользователю
      description: Возвращает список транзакций для указанного пользователя
      operationId: getByUser
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: page
          in: query
          required: false
          schema:
            type: integer
            format: int32
            default: 1
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            format: int32
            default: 10
        - name: sort
          in: query
          required: false
          schema:
            type: string
            default: date
        - name: direction
          in: query
          required: false
          schema:
            type: string
            default: desc
      responses:
        '200':
          description: Список транзакций пользователя
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/PagedModelTransactionDto"
        '403':
          description: Доступ запрещен
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/PagedModelTransactionDto"
      security:
        - bearerAuth: []
...
components:
  schemas:
    UserDto:
      type: object
      description: DTO для представления пользователя
      properties:
        id:
          type: integer
          format: int64
          description: ID пользователя
        username:
          type: string
          description: Имя пользователя
        role:
          type: string
          description: Роль пользователя
          enum:
            - USER
            - ADMIN
        enabled:
          type: boolean
          description: Флаг активности пользователя
    CreateUserRequest:
      type: object
      description: Запрос на создание пользователя
      properties:
        username:
          type: string
          description: Имя пользователя
          example: user
          maxLength: 50
          minLength: 0
        password:
          type: string
          description: Пароль
          example: user
          minLength: 1
        role:
          type: string
          description: Роль пользователя
          enum:
            - USER
            - ADMIN
          example: USER
      required:
        - password
        - role
        - username
    CardDto:
      type: object
      description: DTO для представления банковской карты
      properties:
        id:
          type: integer
          format: int64
          description: ID карты
        number:
          type: string
          description: Номер карты
        holder:
          type: string
          description: Держатель карты
        expiryDate:
          type: string
          format: date
          description: Дата окончания действия
        status:
          type: string
          description: Статус карты
          enum:
            - ACTIVE
            - BLOCKED
            - EXPIRED
        balance:
          type: number
          description: Баланс карты
        deleted:
          type: boolean
          description: Флаг удаления карты
...
    TransactionDto:
      type: object
      description: DTO для представления транзакции
      properties:
        id:
          type: integer
          format: int64
          description: ID транзакции
        from:
          type: integer
          format: int64
          description: ID карты отправителя
        to:
          type: integer
          format: int64
          description: ID карты получателя
        amount:
          type: number
          description: Сумма транзакции
        comment:
          type: string
          description: Комментарий к транзакции
        date:
          type: string
          format: date-time
          description: Дата и время транзакции
...
    PageMetadata:
      type: object
      properties:
        size:
          type: integer
          format: int64
        number:
          type: integer
          format: int64
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int64
    PagedModelUserDto:
      type: object
      properties:
        content:
          type: array
          items:
            "$ref": "#/components/schemas/UserDto"
        page:
          "$ref": "#/components/schemas/PageMetadata"
    PagedModelCardDto:
      type: object
      properties:
        content:
          type: array
          items:
            "$ref": "#/components/schemas/CardDto"
        page:
          "$ref": "#/components/schemas/PageMetadata"
    PagedModelTransactionDto:
      type: object
      properties:
        content:
          type: array
          items:
            "$ref": "#/components/schemas/TransactionDto"
        page:
          "$ref": "#/components/schemas/PageMetadata"
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT